fix: Split mocks and tests
